Blindado programado

+++ TP3-PG/TP3-PG/Blindado.cpp
@@ -1,5 +1,5 @@
 #include "Blindado.h"
-Blindado::Blindado(float _x, float _y, float _w, float _h, float _anchoPantalla, float _altoPantalla, int _vida) :Enemigo(_x, _y, _w, _h, _anchoPantalla, _altoPantalla, _vida)
+Blindado::Blindado(float _x, float _y, float _w, float _h, float _anchoPantalla, float _altoPantalla, float _vida) :Enemigo(_x, _y, _w, _h, _anchoPantalla, _altoPantalla, _vida)
 {
 	bitmapBlindado = NULL;
 }
