+++ TP3-PG/TP3-PG/Cazador.cpp
@@ -5,7 +5,7 @@
 Cazador::Cazador():Enemigo()
 {
 }
-Cazador::Cazador(float _x, float _y, float _w, float _h, float _anchoPantalla, float _altoPantalla, int _vida) : Enemigo(_x, _y, _w, _h, _anchoPantalla, _altoPantalla, _vida)
+Cazador::Cazador(float _x, float _y, float _w, float _h, float _anchoPantalla, float _altoPantalla, float _vida) : Enemigo(_x, _y, _w, _h, _anchoPantalla, _altoPantalla, _vida)
 {
 }
 Cazador::~Cazador()
